requirements.txt
langchain-nvidia-ai-endpoints
python-dotenv
streamlit
langchain
langchain-community
unstructured
faiss-cpu
transformers
